requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.0
cycler==0.12.1
dill==0.3.6
dlib==19.24.1
docker-pycreds==0.4.0
exceptiongroup==1.3.0
face-recognition==1.3.0
face_alignment==1.3.4
face_recognition_models==0.3.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.50.0
fsspec==2025.7.0
ftfy==6.2.0
gdown==3.12.2
gitdb==4.0.11
GitPython==3.1.42
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.6
ifaddr==0.2.0
imageio==2.34.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.2.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
networkx==3.2.1
ninja==1.11.1.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.11.0.86
opencv-python==4.7.0.72
opt_einsum==3.4.0
orjson==3.11.0
packaging==24.0
pandas==2.0.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.25.3
psutil==5.9.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==1.43.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.1
threadpoolctl==3.4.0
tifffile==2024.2.12
tomlkit==0.13.3
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.66.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.35.0
uvloop==0.21.0
wandb==0.15.10
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
zeroconf==0.147.0
